--- src/modals/SendModals/Script/BuildTxModalContent.tsx
@@ -1,79 +1,0 @@
-/*
-Copyright 2018 - 2022 The Alephium Authors
-This file is part of the alephium project.
-
-The library is free software: you can redistribute it and/or modify
-it under the terms of the GNU Lesser General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
-
-The library is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-GNU Lesser General Public License for more details.
-
-You should have received a copy of the GNU Lesser General Public License
-along with the library. If not, see <http://www.gnu.org/licenses/>.
-*/
-
-import { convertAlphToSet } from '@alephium/sdk'
-
-import { ScriptTxData } from '../../../types/transactions'
-import { isAmountWithinRange } from '../../../utils/transactions'
-import { ModalContent, PartialTxData, SendTxModalFooterButtons, useBuildTxCommon, useBytecode } from '../utils'
-
-export interface ScriptBuildTxModalContentProps {
-  data: PartialTxData<ScriptTxData, 'fromAddress'>
-  onSubmit: (data: ScriptTxData) => void
-  onCancel: () => void
-}
-
-const ScriptBuildTxModalContent = ({ data, onSubmit, onCancel }: ScriptBuildTxModalContentProps) => {
-  const [
-    fromAddress,
-    FromAddressSelect,
-    alphAmount,
-    AlphAmountInput,
-    gasAmount,
-    gasPrice,
-    GasSettingsExpandableSection,
-    isCommonReady
-  ] = useBuildTxCommon(data.fromAddress, data.alphAmount, data.gasAmount, data.gasPrice)
-  const [bytecode, BytecodeInput] = useBytecode(data.bytecode ?? '')
-
-  if (fromAddress === undefined) {
-    onCancel()
-    return <></>
-  }
-
-  const isSubmitButtonActive =
-    isCommonReady &&
-    bytecode &&
-    (!alphAmount || isAmountWithinRange(convertAlphToSet(alphAmount), fromAddress.availableBalance))
-
-  return (
-    <>
-      <ModalContent>
-        {FromAddressSelect}
-        {AlphAmountInput}
-        {BytecodeInput}
-      </ModalContent>
-      {GasSettingsExpandableSection}
-      <SendTxModalFooterButtons
-        onSubmit={() =>
-          onSubmit({
-            fromAddress: data.fromAddress,
-            bytecode: bytecode,
-            alphAmount: alphAmount ? alphAmount : undefined,
-            gasAmount: gasAmount.parsed,
-            gasPrice: gasPrice.parsed
-          })
-        }
-        onCancel={onCancel}
-        isSubmitButtonActive={isSubmitButtonActive}
-      />
-    </>
-  )
-}
-
-export default ScriptBuildTxModalContent
