+++ src/pages/Wallet/WalletLayout.tsx
@@ -35,13 +35,12 @@ import Spinner from '../../components/Spinner'
 import { useAddressesContext } from '../../contexts/addresses'
 import { useGlobalContext } from '../../contexts/global'
 import { useSendModalContext } from '../../contexts/sendModal'
-import useDappTxData from '../../hooks/useDappTxData'
 import LogoDarkSrc from '../../images/alephium_logo_dark.svg'
 import LogoLightSrc from '../../images/alephium_logo_light.svg'
 import CenteredModal from '../../modals/CenteredModal'
-import DeployContractSendModal from '../../modals/SendModals/DeployContract/Modal'
-import ScriptSendModal from '../../modals/SendModals/Script/Modal'
-import TransferSendModal from '../../modals/SendModals/Transfer/Modal'
+import SendModalDeployContract from '../../modals/SendModals/SendModalDeployContract'
+import SendModalScript from '../../modals/SendModals/SendModalScript'
+import SendModalTransfer from '../../modals/SendModals/SendModalTransfer'
 import { appHeaderHeightPx, deviceBreakPoints, walletSidebarWidthPx } from '../../style/globalStyles'
 import { TxType } from '../../types/transactions'
 
@@ -56,10 +55,9 @@ const Storage = getStorage()
 
 const WalletLayout: FC = ({ children }) => {
   const { wallet, lockWallet, activeWalletName, login, networkStatus } = useGlobalContext()
-  const { isSendModalOpen, openSendModal, closeSendModal, txType } = useSendModalContext()
+  const { isSendModalOpen, openSendModal, txType } = useSendModalContext()
   const [isPasswordModalOpen, setIsPasswordModalOpen] = useState(false)
   const { refreshAddressesData, isLoadingData } = useAddressesContext()
-  const txData = useDappTxData()
   const navigate = useNavigate()
   const location = useLocation()
   const theme = useTheme()
@@ -143,15 +141,9 @@ const WalletLayout: FC = ({ children }) => {
         </WalletActions>
       </WalletSidebar>
       <AnimatePresence exitBeforeEnter initial={true}>
-        {isSendModalOpen && txType === TxType.TRANSFER && (
-          <TransferSendModal initialTxData={txData} onClose={closeSendModal} />
-        )}
-        {isSendModalOpen && txType === TxType.DEPLOY_CONTRACT && (
-          <DeployContractSendModal initialTxData={txData} onClose={closeSendModal} />
-        )}
-        {isSendModalOpen && txType === TxType.SCRIPT && (
-          <ScriptSendModal initialTxData={txData} onClose={closeSendModal} />
-        )}
+        {isSendModalOpen && txType === TxType.TRANSFER && <SendModalTransfer />}
+        {isSendModalOpen && txType === TxType.DEPLOY_CONTRACT && <SendModalDeployContract />}
+        {isSendModalOpen && txType === TxType.SCRIPT && <SendModalScript />}
         {isPasswordModalOpen && (
           <CenteredModal narrow title="Enter password" onClose={() => setIsPasswordModalOpen(false)}>
             <PasswordConfirmation
